APIMF-1996: add apiKey type validations as warnings
