+++ amf-client/shared/src/test/resources/validations/oas2/security/invalid-apikey-type.json
@@ -1,0 +1,13 @@
+{
+  "swagger": "2.0",
+  "info": {
+    "version": "1",
+    "title": "test"
+  },
+  "paths": {},
+  "securityDefinitions": {
+    "Bearer": {
+      "type": "apiKey"
+    }
+  }
+}
