+++ amf-client/shared/src/test/resources/validations/reports/oas2/invalid-apikey-type.report
@@ -1,0 +1,22 @@
+Model: file://amf-client/shared/src/test/resources/validations/oas2/security/invalid-apikey-type.json
+Profile: OAS 2.0
+Conforms? true
+Number of results: 2
+
+Level: Warning
+
+- Source: http://a.ml/vocabularies/amf/aml#closed-shape-warning
+  Message: Property 'in' is required in a OAS 2.0 apiKey node
+  Level: Warning
+  Target: file://amf-client/shared/src/test/resources/validations/oas2/security/invalid-apikey-type.json#/declarations/securitySchemes/Bearer
+  Property: 
+  Position: Some(LexicalInformation([(9,14)-(11,5)]))
+  Location: file://amf-client/shared/src/test/resources/validations/oas2/security/invalid-apikey-type.json
+
+- Source: http://a.ml/vocabularies/amf/aml#closed-shape-warning
+  Message: Property 'name' is required in a OAS 2.0 apiKey node
+  Level: Warning
+  Target: file://amf-client/shared/src/test/resources/validations/oas2/security/invalid-apikey-type.json#/declarations/securitySchemes/Bearer
+  Property: 
+  Position: Some(LexicalInformation([(9,14)-(11,5)]))
+  Location: file://amf-client/shared/src/test/resources/validations/oas2/security/invalid-apikey-type.json
